[ refactoring ] install script renewal
part 1 of *several*

each create_config script is now responsible for installing the software
it is depending on plus linking/copying the config files etc.

+++ pass/create_config.bash
@@ -1,19 +1,32 @@
 #!/bin/bash
 # create_config for -- <<PROGNAME>>
 
-prgname="rofipassmenu"
+prgname="pass"
 
 echo -e "\e[1;36mInstalling ... ${prgname} ... configuration files ...\e[0m"
 sleep 1
 
 source INSTALL_ALL/config.bash
-source INSTALL_ALL/yay.bash
-source INSTALL_ALL/pacman.bash
 
 main()
 {
-    rm -rf ${HOME}/.password-store
-    ln -sfv ${HOME}/ownCloud/Documents/pass ${HOME}/.password-store
+    installation
+    configuration
+}
+
+installation()
+{
+    if [[ $(check_installed ${prgname}) -eq 0 ]] ; then
+        sudo pacman -S ${prgname}
+    fi
+}
+
+configuration()
+{
+    if [[ $(check_symlink ~/.password-store) == "false" ]] ; then
+        rm -rf ${HOME}/.password-store
+        ln -sfv ${HOME}/ownCloud/Documents/pass ${HOME}/.password-store
+    fi
 }
 
 main
